+++ crates/meroctl/src/cli.rs
@@ -1,10 +1,11 @@
 use clap::{Parser, Subcommand};
 
-use crate::config as struct_config;
+use crate::defaults;
 
 mod config;
 mod init;
 mod run;
+
 #[derive(Debug, Parser)]
 #[command(author, about, version)]
 pub struct RootCommand {
@@ -25,7 +26,7 @@ pub enum SubCommands {
 #[derive(Debug, Parser)]
 pub struct RootArgs {
     /// Directory for config and data
-    #[arg(long, value_name = "PATH", default_value_t = struct_config::default_chat_dir())]
+    #[arg(long, value_name = "PATH", default_value_t = defaults::default_chat_dir())]
     #[arg(env = "CALIMERO_HOME", hide_env_values = true)]
     pub home: camino::Utf8PathBuf,
 
@@ -37,9 +38,9 @@ pub struct RootArgs {
 impl RootCommand {
     pub async fn run(self) -> eyre::Result<()> {
         match self.action {
-            SubCommands::Init(init) => return init.run(self.args),
-            SubCommands::Config(config) => return config.run(self.args),
-            SubCommands::Run(run) => return run.run(self.args).await,
+            SubCommands::Init(init) => init.run(self.args),
+            SubCommands::Config(config) => config.run(self.args),
+            SubCommands::Run(run) => run.run(self.args).await,
         }
     }
 }
